# docs/reference/openapi/aegis-v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Aegis API",
    "description": "Centralized, explainable authorization API.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/stores/{storeId}/assertions/{authorizationModelId}": {
      "get": {
        "tags": [
          "StoreAssertions"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorizationModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AegisCompatReadAssertionsResponseDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisCompatReadAssertionsResponseDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisCompatReadAssertionsResponseDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "StoreAssertions"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorizationModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AegisCompatWriteAssertionsRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AegisCompatWriteAssertionsRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AegisCompatWriteAssertionsRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StringApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{storeId}/assertions/{authorizationModelId}/run": {
      "post": {
        "tags": [
          "StoreAssertions"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorizationModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{storeId}/assertions/{authorizationModelId}/runs": {
      "get": {
        "tags": [
          "StoreAssertions"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorizationModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunListResponseDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunListResponseDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunListResponseDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{storeId}/assertions/runs/{runId}": {
      "get": {
        "tags": [
          "StoreAssertions"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisAssertionRunRecordDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/stores/{storeId}/assertions/{authorizationModelId}/generate-from-audit": {
      "post": {
        "tags": [
          "StoreAssertions"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorizationModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AegisGenerateAssertionsFromAuditRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AegisGenerateAssertionsFromAuditRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AegisGenerateAssertionsFromAuditRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AegisGenerateAssertionsFromAuditResponseDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisGenerateAssertionsFromAuditResponseDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AegisGenerateAssertionsFromAuditResponseDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stores/{storeId}": {
      "get": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDtoApiResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Stores"
        ],
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StringApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StringApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AegisApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AegisAssertionRunListResponseDto": {
        "type": "object",
        "properties": {
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AegisAssertionRunRecordDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AegisAssertionRunListResponseDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AegisAssertionRunListResponseDto"
          },
          "error": {
            "$ref": "#/components/schemas/AegisApiError"
          }
        },
        "additionalProperties": false
      },
      "AegisAssertionRunRecordDto": {
        "type": "object",
        "properties": {
          "run_id": {
            "type": "string",
            "nullable": true
          },
          "store_id": {
            "type": "string",
            "nullable": true
          },
          "authorization_model_id": {
            "type": "string",
            "nullable": true
          },
          "started_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "$ref": "#/components/schemas/AegisAssertionRunSummaryDto"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AegisAssertionRunResultItemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AegisAssertionRunRecordDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AegisAssertionRunRecordDto"
          },
          "error": {
            "$ref": "#/components/schemas/AegisApiError"
          }
        },
        "additionalProperties": false
      },
      "AegisAssertionRunResultItemDto": {
        "type": "object",
        "properties": {
          "tuple_key": {
            "$ref": "#/components/schemas/AegisCompatTupleKeyDto"
          },
          "expected": {
            "type": "boolean"
          },
          "actual": {
            "type": "boolean"
          },
          "passed": {
            "type": "boolean"
          },
          "decision": {
            "type": "string",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "explain_trace_id": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AegisAssertionRunSummaryDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "passed": {
            "type": "integer",
            "format": "int32"
          },
          "failed": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "AegisCompatAssertionDto": {
        "type": "object",
        "properties": {
          "tuple_key": {
            "$ref": "#/components/schemas/AegisCompatTupleKeyDto"
          },
          "expectation": {
            "type": "boolean"
          },
          "contextual_tuples": {
            "$ref": "#/components/schemas/AegisCompatContextualTuplesDto"
          }
        },
        "additionalProperties": false
      },
...
      "AegisCompatContextualTuplesDto": {
        "type": "object",
        "properties": {
          "tuple_keys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AegisCompatTupleKeyDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AegisCompatReadAssertionsResponseDto": {
        "type": "object",
        "properties": {
          "authorization_model_id": {
            "type": "string",
            "nullable": true
          },
          "assertions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AegisCompatAssertionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AegisCompatReadAssertionsResponseDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AegisCompatReadAssertionsResponseDto"
          },
          "error": {
            "$ref": "#/components/schemas/AegisApiError"
          }
        },
        "additionalProperties": false
      },
      "AegisCompatTupleKeyDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "nullable": true
          },
          "relation": {
            "type": "string",
            "nullable": true
          },
          "object": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AegisCompatWriteAssertionsRequestDto": {
        "type": "object",
        "properties": {
          "assertions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AegisCompatAssertionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AegisGenerateAssertionsFromAuditRequestDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "nullable": true
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "append": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "AegisGenerateAssertionsFromAuditResponseDto": {
        "type": "object",
        "properties": {
          "authorization_model_id": {
            "type": "string",
            "nullable": true
          },
          "generated_count": {
            "type": "integer",
            "format": "int32"
          },
          "appended": {
            "type": "boolean"
          },
          "assertions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AegisCompatAssertionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AegisGenerateAssertionsFromAuditResponseDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/AegisGenerateAssertionsFromAuditResponseDto"
          },
          "error": {
            "$ref": "#/components/schemas/AegisApiError"
          }
        },
        "additionalProperties": false
      },
...
      "StoreDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "modelCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "relationshipCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tenantId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "StoreDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/StoreDto"
          },
          "error": {
            "$ref": "#/components/schemas/AegisApiError"
          }
        },
        "additionalProperties": false
      },
...
      "StringApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "string",
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/AegisApiError"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}